refactor: accounts service initialization
