+++ packages/interledger/src/connector/core/middleware/auth.ts
@@ -1,4 +1,4 @@
-import { IlpAccount } from '../../../accounts'
+import { IlpAccount } from '../../../accounts/types'
 
 export interface AuthState {
   account?: IlpAccount
